requirements.txt
Babel==2.16.0
Kvsqlite==0.2.1
Pillow==10.4.0
pronotepy==2.13.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
pytz==2024.2
pyzbar==0.1.9
rich==13.8.1
